+++ types/components/Datatable.d.ts
@@ -95,4 +95,8 @@ export default class {
      * }
      */
     requestErrorHook(code: number, xhr: XMLHttpRequest): void;
+    /**
+     * Select the column found by a the column selector
+     */
+    column(columnSelector: any, modifier?: DataTables.ObjectSelectorModifier): DataTables.ColumnMethods;
 }